pyproject.toml
[project]
name = "brf-helper"
version = "0.1.0"
description = "AI-powered Swedish BRF annual report analysis tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "pypdf>=5.1.0",
    "google-generativeai>=0.8.3",
    "chromadb>=0.5.23",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
